refactor: Consolidate community components into a single CommunityPage
- Merged multiple community-related components (ChatTab, CommunityMain, MembersTab, etc.) into a single CommunityPage for improved structure and maintainability.
- Enhanced loading and profile completion UI with a more cohesive design.
- Streamlined message handling and user presence management, ensuring a smoother user experience.
- Removed redundant components and optimized state management within the useCommunity hook.
